# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaravelAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/saveEmpresa": {
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Guardar Empresa",
                "description": "Guardar Empresa en la db",
                "operationId": "saveEmpresa",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Process Succesfull"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updateEmpresas": {
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar Empresas",
                "description": "Actualizar empresa en la db",
                "operationId": "updateEmpresa",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Process Succesfull"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/removerEmpresa": {
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Eliminar Empresa",
                "description": "Eliminar Empresa en la db",
                "operationId": "removeEmpresa",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Process Succesfull"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Get Empresas",
                "description": "Obtiene todas las empresas",
                "operationId": "getEmpresas",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "razonSocial": {
                                                "type": "string",
                                                "example": "Dimension Comercial S.A"
                                            },
                                            "cliente a Cargo": {
                                                "type": "string",
                                                "example": "Luis Ventocilla"
                                            },
                                            "created": {
                                                "type": "string",
                                                "example": "2020-09-17 03:21:05"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}